Fabric port

+++ src/main/java/com/natim6/lazyengines/mixin/BoilerDataMixin.java
@@ -74,7 +74,7 @@ public class BoilerDataMixin {
     }
 
     @Shadow
-    public float waterSupply;
+    public double waterSupply;
     @Final
     @Shadow
     private static int waterSupplyPerLevel;
